fix(Endpoint): Move endpoint used only source code
The "cipher","connectivity", "https", and the "text_serializer" are
only serving for endpoint. Moving these files under endpoint may make
more sense.

Close #638
