+++ endpoint/BUILD
@@ -1,3 +1,5 @@
+package(default_visibility = ["//visibility:public"])
+
 cc_library(
     name = "endpoint_core",
     srcs = ["endpoint_core.c"],
@@ -11,9 +13,9 @@ cc_library(
     deps = [
         "//common:ta_errors",
         "//common:ta_logger",
-        "//utils:cipher",
-        "//utils:https",
-        "//utils:text_serializer",
+        "//endpoint:cipher",
+        "//endpoint:https",
+        "//endpoint:text_serializer",
         "//utils:tryte_byte_conv",
     ],
 )
@@ -33,9 +35,9 @@ cc_binary(
     deps = [
         "//common:ta_errors",
         "//common:ta_logger",
-        "//utils:cipher",
-        "//utils:https",
-        "//utils:text_serializer",
+        "//endpoint:cipher",
+        "//endpoint:https",
+        "//endpoint:text_serializer",
         "//utils:tryte_byte_conv",
     ],
 )
@@ -47,5 +49,45 @@ platform(
         # FIXME: The generic CPU name will be replaced in the future
         # ref: https://github.com/bazelbuild/platforms/blob/master/cpu/BUILD
         "@platforms//cpu:arm",
+    ],
+)
+
+cc_library(
+    name = "cipher",
+    srcs = ["cipher.c"],
+    hdrs = ["cipher.h"],
+    deps = [
+        "//common:ta_errors",
+        "//common:ta_logger",
+        "@mbedtls_2_16_6",
+    ],
+)
+
+cc_library(
+    name = "https",
+    srcs = ["https.c"],
+    hdrs = ["https.h"],
+    defines = select({
+        ":https_enable": ["ENDPOINT_HTTPS"],
+        "//conditions:default": [],
+    }),
+    deps = [
+        "//common:ta_errors",
+        "//endpoint/connectivity:conn_http",
+    ],
+)
+
+config_setting(
+    name = "https_enable",
+    values = {"define": "https=enable"},
+)
+
+cc_library(
+    name = "text_serializer",
+    srcs = ["text_serializer.c"],
+    hdrs = ["text_serializer.h"],
+    deps = [
+        ":cipher",
+        "//common:ta_errors",
     ],
 )
