+++ endpoint/connectivity/BUILD
