+++ endpoint/unit-test/BUILD
@@ -4,8 +4,8 @@ cc_test(
         "test_http.c",
     ],
     deps = [
+        "//endpoint/connectivity:conn_http",
         "//tests:test_define",
-        "//utils/connectivity:conn_http",
     ],
 )
